requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Cloud & BigQuery
google-cloud-bigquery==3.13.0
google-auth==2.25.2
google-api-python-client==2.108.0
db-dtypes==1.5.0  # Required for BigQuery data type conversions

# Data processing
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
prophet==1.1.5
arch==6.2.0  # For GARCH models
statsmodels==0.14.0

# Time series & forecasting
pmdarima==2.0.4

# Explainability
shap==0.43.0

# Real-time & Caching
redis==5.0.1
kafka-python==2.0.2

# Notifications
twilio==8.10.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Logging & Monitoring
structlog==23.2.0

# Type checking (optional - only needed for mypy)
# mypy==1.7.1
# types-requests==2.31.0.10  # Install separately if needed: pip install types-requests

